# src/main/resources/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger DoorKeeper
  license:
    name: MIT
servers:
  - url: http://localhost:8080/
paths:
...
  /positions/{id}:
    get:
      summary: Get representation of position
      operationId: getPosition
      tags:
        - office
      parameters:
        - in: header
          name: X-Token
          schema:
            type: string
          required: true
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Done
          content:
            image/png:
              schema:
                type: string
                format: byte
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /positions:
    post:
      summary: Set positions by uploading map
      operationId: setPositions
      tags:
        - office
      parameters:
        - in: header
          name: X-Token
          schema:
            type: string
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                officeMap:
                  type: string
                  format: binary
      responses:
        '200':
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    PositionsResponse:
      type: object
      required:
        - numberOfPositions
      properties:
        message:
          type: number
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string